Canvas layout: fix double value

+++ src/lib/edje/edje_private.h
@@ -2292,7 +2292,7 @@ typedef struct
   union
   {
     int ni; // number integer
-    int nd; // number double
+    double nd; // number double
     Efl_Text_Format_Wrap wrap;
     Efl_Text_Style_Backing_Type backing;
     Efl_Text_Style_Underline_Type underline;
